chore(suite-native): connect&unlock redirect
