+++ suite-native/module-authorize-device/src/screens/connect/ConnectAndUnlockDeviceScreen.tsx
@@ -1,12 +1,8 @@
-import { useCallback, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-
-import { useIsFocused } from '@react-navigation/native';
 
 import { bluetoothActions } from '@suite-common/bluetooth';
 import {
     selectIsBluetoothSupportedByDevice,
-    selectIsDeviceAuthorized,
     selectIsDeviceConnected,
 } from '@suite-common/wallet-core';
 import { ConnectAndUnlockDeviceScreenContent } from '@suite-native/device';
@@ -15,11 +11,10 @@ import {
     AuthorizeDeviceStackParamList,
     AuthorizeDeviceStackRoutes,
     RootStackParamList,
-    Screen,
     StackToStackCompositeScreenProps,
 } from '@suite-native/navigation';
 
-import { ConnectDeviceScreenHeader } from '../../components/connect/ConnectDeviceScreenHeader';
+import { ConnectDeviceScreen } from '../../components/connect/ConnectDeviceScreen';
 
 export const ConnectAndUnlockDeviceScreen = ({
     navigation,
@@ -31,20 +26,11 @@ export const ConnectAndUnlockDeviceScreen = ({
     const dispatch = useDispatch();
 
     const isDeviceConnected = useSelector(selectIsDeviceConnected);
-    const isDeviceAuthorized = useSelector(selectIsDeviceAuthorized);
 
     const isBluetoothEnabled = useFeatureFlag(FeatureFlag.IsBluetoothEnabled);
     const isBluetoothSupportedByDevice = useSelector(selectIsBluetoothSupportedByDevice);
     const isBluetoothButtonVisible =
         isBluetoothEnabled && (!isDeviceConnected || isBluetoothSupportedByDevice);
-
-    const isFocused = useIsFocused();
-
-    const navigateBack = useCallback(() => {
-        if (navigation.canGoBack()) {
-            navigation.goBack();
-        }
-    }, [navigation]);
 
     const navigateToTurnOnAndUnlockDeviceScreen = () => {
         // Make sure auto-connect is enabled in case some device was manually disconnected.
@@ -52,34 +38,13 @@ export const ConnectAndUnlockDeviceScreen = ({
         navigation.replace(AuthorizeDeviceStackRoutes.TurnOnAndUnlockDevice);
     };
 
-    useEffect(() => {
-        if (!isFocused || !isDeviceConnected) return;
-
-        if (isDeviceAuthorized) {
-            // When selected device become connected, we need to navigate out of this screen.
-            navigateBack();
-        } else {
-            console.warn(' == meow == authorize device thnk needs to be replaced here ');
-            // If user cancelled the authorization, we need to authorize the device again.
-            // requestPrioritizedDeviceAccess({
-            //     deviceCallback: () => dispatch(authorizeDeviceThunk()),
-            // });
-        }
-    }, [isDeviceAuthorized, isDeviceConnected, isFocused, navigateBack]);
-
     return (
-        <Screen
-            header={<ConnectDeviceScreenHeader />}
-            noHorizontalPadding
-            noBottomPadding
-            hasBottomInset={false}
-            isScrollable={false}
-        >
+        <ConnectDeviceScreen>
             <ConnectAndUnlockDeviceScreenContent
                 onConnectViaBluetooth={
                     isBluetoothButtonVisible ? navigateToTurnOnAndUnlockDeviceScreen : undefined
                 }
             />
-        </Screen>
+        </ConnectDeviceScreen>
     );
 };
